requirements.txt
# Claude API
anthropic>=0.40.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0

# Database
# sqlite3 is built-in to Python

# API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Environment / config
python-dotenv>=1.0.0
